Cleanup of similar OutputBuffers
